+++ frontend/src/app/state/playlist/rulesTree/rulesTree.actions.ts
@@ -1,11 +1,12 @@
 import * as React from 'react';
 import { PlaylistAction } from '../playlist.reducer';
-import { RuleOrRuleGroup } from '~/app/types/rule';
+import { Rule, RuleOrRuleGroup } from '~/app/types/rule';
 import { createRule, createRuleGroup } from '~/app/types/rule.factory';
 import { TreeNodeData } from '~/common/components/TreeView/types/tree';
 
 export type RulesTreeActions = {
     addRule(node: TreeNodeData): void;
+    addRuleEntity(node: TreeNodeData, entity: Rule): void;
     addRuleGroup(node: TreeNodeData): void;
     removeRule(node: TreeNodeData): void;
     changeExpand(node: TreeNodeData<RuleOrRuleGroup>, isExpanded: boolean): void;
@@ -19,6 +20,13 @@ export const createRulesTreeActions = (
             dispatcher({
                 type: 'ruleTree:add',
                 entity: createRule(),
+                treeNode: node,
+            });
+        },
+        addRuleEntity: (node: TreeNodeData, entity: Rule) => {
+            dispatcher({
+                type: 'ruleTree:add',
+                entity: entity,
                 treeNode: node,
             });
         },