+++ frontend/src/common/components/TreeView/TreeView.tsx
@@ -2,6 +2,10 @@ import * as React from 'react';
 import { Override, useOverrides } from '~/common/hooks/useOverrides';
 import { TreeNode, TreeNodeProps } from '~/common/components/TreeView/TreeNode';
 import { TreeNodeData, TreeViewData } from '~/common/components/TreeView/types/tree';
+import { Button } from '../Button';
+import { Icon } from '../Icon';
+import { TreeViewMultiAdd } from './TreeViewMultiAdd';
+import { PlaylistContext } from '~/app/state/playlist/playlist.context';
 
 type TreeViewProps = {
     data: TreeViewData;
@@ -13,6 +17,9 @@ type TreeViewProps = {
 };
 
 export const TreeView: React.FC<TreeViewProps> = props => {
+    const playlistContext = React.useContext(PlaylistContext);
+    let [showMultiAdd, setShowMultiAdd] = React.useState(false);
+
     const [Node, nodeProps] = useOverrides(props.overrides && props.overrides.Node, TreeNode);
 
     const { data } = props;
@@ -30,6 +37,26 @@ export const TreeView: React.FC<TreeViewProps> = props => {
 
     return (
         <>
+            <div className='multi-add'>
+                <Button onClick={_ => setShowMultiAdd(true)}>
+                    <Icon type="library_add" />
+                </Button>
+            </div>
+            <div>
+                {showMultiAdd && (
+                    <TreeViewMultiAdd
+                        onClose={() => setShowMultiAdd(false)}
+                        onConfirm={(rules) => {
+                            var node = playlistContext.getLastTreeNode();
+                            if (node)
+                                for (let i = rules.length - 1; i >= 0; i--) {
+                                    playlistContext.addRuleEntity(node, rules[i]);
+                                }
+                            setShowMultiAdd(false);
+                        }}
+                    />
+                )}
+            </div>
             {getRootNodes().map(nodeData => (
                 <Node
                     key={nodeData.id}